Created a file to locate all the elements of compose
